+++ ChangeLog.md
@@ -13,6 +13,7 @@ Other enhancements:
 * `stack upgrade` makes less assumptions about archive format. See
   [#5288](https://github.com/commercialhaskell/stack/issues/5288)
 * Add a `--no-run` flag to the `script` command when compiling.
+* Add GHC installation hooks
 
 Bug fixes: